requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
svglib==1.5.1
reportlab==4.0.7
cairosvg==2.7.1
Pillow==10.1.0
pypdf2==3.0.1
lxml==5.1.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-genai>=1.47.0
python-dotenv>=1.0.0
numpy==1.24.3
rembg
onnxruntime